pyproject.toml
[tool.poetry]
name = "instructionbook"
version = "0.1.0"
description = ""
authors = ["santos borom <santos.borom@arkastra.com>"]

[tool.poetry.dependencies]
python = "^3.9"
plotly = "^4.14.3"
pandas = "^1.2.4"
numpy = "^1.20.3"
altair = "^4.1.0"
streamlit = "^0.82.0"
prophet = "^1.0.1"
sklearn = "^0.0"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
